refactor: separate chat and video view and controllers
