+++ src/view/chat/commands/leaveRoom.ts
@@ -1,6 +1,6 @@
 import { Namespace, Socket } from 'socket.io'
 import { NotificationMessage } from '../../../model/message'
-import { ChatController } from '../../../controllers/chatController'
+import { ChatController } from '../../../controllers/chat/chatController'
 import { chatReaction } from '../utils'
 import { SerializerImpl } from '../../../model/presentation/serialization/messageSerializer'
