+++ src/view/sessionNamespace.ts
@@ -1,0 +1,5 @@
+import { Namespace } from 'socket.io'
+
+export interface SessionNamespace {
+  registerCommands(namespace: Namespace): void
+}
