--- src/view/video/videoCommands.ts
@@ -1,23 +1,0 @@
-// import { Namespace } from "socket.io";
-
-// export async function registerVideoCommands(videoNamespace: Namespace) {
-
-//     // videoController: VideoController = VideoController(videoNamespace)
-
-//     // videoNamespace.on("connection", (socket: Socket) => {
-
-//     //     socket.on("Pause", () => {
-
-//     //     });
-
-//     //     socket.on("Play", () => {
-
-//     //     });
-
-//     //     socket.on("TimelineChange", () => {
-
-//     //     });
-
-//     // });
-
-// }
